# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SBS Agent AI Assistant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateOpenaiConversationBody {
  title: string;
}

export interface SendOpenaiMessageBody {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface Spiel {
  id: number;
  title: string;
  category: string;
  content: string;
  createdAt: string;
}

export interface CreateSpielBody {
  title: string;
  category: string;
  content: string;
}

export interface UpdateSpielBody {
  title?: string;
  category?: string;
  content?: string;
}

export interface OpenaiError {
  error: string;
}