# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SBS Agent AI Assistant API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: openai
    description: OpenAI AI chat and audio operations
  - name: spiels
    description: Saved spiels / response templates
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /openai/conversations:
    get:
      operationId: listOpenaiConversations
      tags: [openai]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiConversation"
    post:
      operationId: createOpenaiConversation
      tags: [openai]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOpenaiConversationBody"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversation"

  /openai/conversations/{id}:
    get:
      operationId: getOpenaiConversation
      tags: [openai]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
    delete:
      operationId: deleteOpenaiConversation
      tags: [openai]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"

  /openai/conversations/{id}/messages:
    get:
      operationId: listOpenaiMessages
      tags: [openai]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiMessage"
    post:
      operationId: sendOpenaiMessage
      tags: [openai]
      summary: Send a customer message and receive a streaming SBS spiel response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOpenaiMessageBody"
      responses:
        "200":
          description: SSE stream of assistant text chunks
          content:
            text/event-stream: {}

  /spiels:
    get:
      operationId: listSpiels
      tags: [spiels]
      summary: List all saved spiels
      responses:
        "200":
          description: List of spiels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Spiel"
    post:
      operationId: createSpiel
      tags: [spiels]
      summary: Save a new spiel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSpielBody"
      responses:
        "201":
          description: Created spiel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Spiel"

  /spiels/{id}:
    put:
      operationId: updateSpiel
      tags: [spiels]
      summary: Update a saved spiel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSpielBody"
      responses:
        "200":
          description: Updated spiel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Spiel"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
    delete:
      operationId: deleteSpiel
      tags: [spiels]
      summary: Delete a saved spiel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    OpenaiConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt

    OpenaiMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt

    CreateOpenaiConversationBody:
      type: object
      properties:
        title:
          type: string
      required:
        - title

    SendOpenaiMessageBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content

    OpenaiConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/OpenaiMessage"
      required:
        - id
        - title
        - createdAt
        - messages

    Spiel:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - category
        - content
        - createdAt

    CreateSpielBody:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
        content:
          type: string
      required:
        - title
        - category
        - content

    UpdateSpielBody:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
        content:
          type: string

    OpenaiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error