# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SBS Agent AI Assistant API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all conversations
 */
export const ListOpenaiConversationsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListOpenaiConversationsResponse = zod.array(
  ListOpenaiConversationsResponseItem,
);

/**
 * @summary Create a new conversation
 */
export const CreateOpenaiConversationBody = zod.object({
  title: zod.string(),
});

/**
 * @summary Get conversation with messages
 */
export const GetOpenaiConversationParams = zod.object({
  id: zod.coerce.number(),
});

export const GetOpenaiConversationResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.coerce.date(),
  messages: zod.array(
    zod.object({
      id: zod.number(),
      conversationId: zod.number(),
      role: zod.string(),
      content: zod.string(),
      createdAt: zod.coerce.date(),
    }),
  ),
});

/**
 * @summary Delete a conversation
 */
export const DeleteOpenaiConversationParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List messages in a conversation
 */
export const ListOpenaiMessagesParams = zod.object({
  id: zod.coerce.number(),
});

export const ListOpenaiMessagesResponseItem = zod.object({
  id: zod.number(),
  conversationId: zod.number(),
  role: zod.string(),
  content: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListOpenaiMessagesResponse = zod.array(
  ListOpenaiMessagesResponseItem,
);

/**
 * @summary Send a customer message and receive a streaming SBS spiel response
 */
export const SendOpenaiMessageParams = zod.object({
  id: zod.coerce.number(),
});

export const SendOpenaiMessageBody = zod.object({
  content: zod.string(),
});

/**
 * @summary List all saved spiels
 */
export const ListSpielsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  category: zod.string(),
  content: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListSpielsResponse = zod.array(ListSpielsResponseItem);

/**
 * @summary Save a new spiel
 */
export const CreateSpielBody = zod.object({
  title: zod.string(),
  category: zod.string(),
  content: zod.string(),
});

/**
 * @summary Update a saved spiel
 */
export const UpdateSpielParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateSpielBody = zod.object({
  title: zod.string().optional(),
  category: zod.string().optional(),
  content: zod.string().optional(),
});

export const UpdateSpielResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  category: zod.string(),
  content: zod.string(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete a saved spiel
 */
export const DeleteSpielParams = zod.object({
  id: zod.coerce.number(),
});