# lib/api-zod/src/generated/types/updateSpielBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SBS Agent AI Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateSpielBody {
  title?: string;
  category?: string;
  content?: string;
}